extendr_ffi

Module backports

Backport functions

R’s C API is changing and is stabilizing. As such some functions that were available in previous versions of R are not available in later versions of R, or they cause a warning in R CMD check.

Use the functions in this module to ensure backwards compatibility.

Functions

DATAPTR_RO 🔒 ⚠

R_ClosureBody 🔒 ⚠

R_ClosureEnv 🔒 ⚠

R_ClosureFormals 🔒 ⚠

R_ParentEnv 🔒 ⚠

R_getVar 🔒 ⚠

R_getVarEx 🔒 ⚠

dataptr⚠

Access a DATAPTR

get_closure_body⚠

Return the body of a closure

get_closure_env⚠

Return the environment of a closure

get_closure_formals⚠

Access a closure’s arguments

get_parent_env⚠

Returns the enclosing environment of env, which will usually be of type ENVSXP, except for the special environment R_EmptyEnv, which terminates the environment chain; its enclosing environment is R_NilValue.

get_var⚠

Returns a variable from an environment

get_var_in_frame⚠

Returns the value of the requested variable in an environment